threads: add reply feature

+++ database/migrations/2019_08_02_142904_create_replies_table.php
@@ -15,8 +15,8 @@ class CreateRepliesTable extends Migration
     {
         Schema::create('replies', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('thread_id');
-            $table->integer('user_id');
+            $table->bigInteger('thread_id');
+            $table->bigInteger('user_id');
             $table->text('body');
             $table->timestamps();
         });
